feat: upgrade boardgame.io and handle both new and deprecated storage class methods
This upgrades the minimum boardgame.io version to 0.41.1 and supports 
storage connectors that use both the deprecated and new method names.

BREAKING CHANGE: Versions of boardgame.io <0.41.1 are no longer 
supported and other uses that directly relied on calling `listGames` or 
`createGame` will need to be updated to use the new `listMatches` and 
`createMatch` method names.

Fix #3